+++ src/functions/index.ts
@@ -43,6 +43,7 @@ import {
   SecretFetchForGoogleAccessToken,
   SecretFetchForGoogleSecretManager,
 } from './secret/index.js';
+import { TypeScriptGetDecoratorRendererForGoogleSpanner } from './typescript/index.js';
 
 export function registerFunctions(context: ModuleRegistrationContext) {
   context.registerFunctionImplementations(
@@ -78,5 +79,6 @@ export function registerFunctions(context: ModuleRegistrationContext) {
     ProjectPushArtefactForCloudFunctions,
     SecretFetchForGoogleAccessToken,
     SecretFetchForGoogleSecretManager,
+    TypeScriptGetDecoratorRendererForGoogleSpanner,
   );
 }